revert PR#5010 to fix proxy startup npe.
