+++ sharding-proxy/sharding-proxy-backend/src/main/java/org/apache/shardingsphere/shardingproxy/backend/schema/impl/EncryptSchema.java
@@ -28,10 +28,17 @@ import org.apache.shardingsphere.encrypt.metadata.EncryptTableMetaDataDecorator;
 import org.apache.shardingsphere.encrypt.rule.EncryptRule;
 import org.apache.shardingsphere.orchestration.core.common.event.EncryptRuleChangedEvent;
 import org.apache.shardingsphere.shardingproxy.backend.schema.LogicSchema;
+import org.apache.shardingsphere.shardingproxy.backend.schema.LogicSchemas;
 import org.apache.shardingsphere.shardingproxy.config.yaml.YamlDataSourceParameter;
+import org.apache.shardingsphere.shardingproxy.context.ShardingProxyContext;
+import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaData;
+import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaDataLoader;
+import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.underlying.common.metadata.ShardingSphereMetaData;
+import org.apache.shardingsphere.underlying.common.metadata.datasource.DataSourceMetas;
 import org.apache.shardingsphere.underlying.common.metadata.decorator.SchemaMetaDataDecorator;
 
+import javax.sql.DataSource;
 import java.sql.SQLException;
 import java.util.Map;
 
@@ -41,6 +48,8 @@ import java.util.Map;
 @Getter
 public final class EncryptSchema extends LogicSchema {
     
+    private final ShardingSphereMetaData metaData;
+    
     private final ShardingRule shardingRule;
     
     private EncryptRule encryptRule;
@@ -49,12 +58,20 @@ public final class EncryptSchema extends LogicSchema {
         super(name, dataSources);
         encryptRule = new EncryptRule(encryptRuleConfiguration);
         shardingRule = new ShardingRule(new ShardingRuleConfiguration(), getDataSources().keySet());
+        metaData = createMetaData();
     }
     
-    @Override
-    public ShardingSphereMetaData getMetaData() {
-        return new ShardingSphereMetaData(getPhysicalMetaData().getDataSources(),
-                SchemaMetaDataDecorator.decorate(getPhysicalMetaData().getSchema(), encryptRule, new EncryptTableMetaDataDecorator()));
+    private ShardingSphereMetaData createMetaData() throws SQLException {
+        DataSourceMetas dataSourceMetas = new DataSourceMetas(LogicSchemas.getInstance().getDatabaseType(), getDatabaseAccessConfigurationMap());
+        SchemaMetaData schemaMetaData = createSchemaMetaData();
+        return new ShardingSphereMetaData(dataSourceMetas, schemaMetaData);
+    }
+    
+    private SchemaMetaData createSchemaMetaData() throws SQLException {
+        DataSource dataSource = getBackendDataSource().getDataSources().values().iterator().next();
+        int maxConnectionsSizePerQuery = ShardingProxyContext.getInstance().getProperties().<Integer>getValue(ConfigurationPropertyKey.MAX_CONNECTIONS_SIZE_PER_QUERY);
+        return SchemaMetaDataDecorator.decorate(SchemaMetaDataLoader.load(dataSource, maxConnectionsSizePerQuery,
+            LogicSchemas.getInstance().getDatabaseType().getName()), encryptRule, new EncryptTableMetaDataDecorator());
     }
     
     /**
