+++ sharding-proxy/sharding-proxy-backend/src/main/java/org/apache/shardingsphere/shardingproxy/backend/schema/impl/ShadowSchema.java
@@ -27,9 +27,16 @@ import org.apache.shardingsphere.core.rule.ShadowRule;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 import org.apache.shardingsphere.orchestration.core.common.event.ShadowRuleChangedEvent;
 import org.apache.shardingsphere.shardingproxy.backend.schema.LogicSchema;
+import org.apache.shardingsphere.shardingproxy.backend.schema.LogicSchemas;
 import org.apache.shardingsphere.shardingproxy.config.yaml.YamlDataSourceParameter;
+import org.apache.shardingsphere.shardingproxy.context.ShardingProxyContext;
+import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaData;
+import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaDataLoader;
+import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.underlying.common.metadata.ShardingSphereMetaData;
+import org.apache.shardingsphere.underlying.common.metadata.datasource.DataSourceMetas;
 
+import javax.sql.DataSource;
 import java.sql.SQLException;
 import java.util.Map;
 
@@ -39,6 +46,8 @@ import java.util.Map;
 @Getter
 public final class ShadowSchema extends LogicSchema {
     
+    private final ShardingSphereMetaData metaData;
+    
     private final ShardingRule shardingRule;
     
     private ShadowRule shadowRule;
@@ -47,11 +56,19 @@ public final class ShadowSchema extends LogicSchema {
         super(name, dataSources);
         shadowRule = new ShadowRule(shadowRuleConfiguration);
         shardingRule = new ShardingRule(new ShardingRuleConfiguration(), getDataSources().keySet());
+        metaData = createMetaData();
     }
-
-    @Override
-    public ShardingSphereMetaData getMetaData() {
-        return getPhysicalMetaData();
+    
+    private ShardingSphereMetaData createMetaData() throws SQLException {
+        DataSourceMetas dataSourceMetas = new DataSourceMetas(LogicSchemas.getInstance().getDatabaseType(), getDatabaseAccessConfigurationMap());
+        SchemaMetaData schemaMetaData = createSchemaMetaData();
+        return new ShardingSphereMetaData(dataSourceMetas, schemaMetaData);
+    }
+    
+    private SchemaMetaData createSchemaMetaData() throws SQLException {
+        DataSource dataSource = getBackendDataSource().getDataSources().values().iterator().next();
+        int maxConnectionsSizePerQuery = ShardingProxyContext.getInstance().getProperties().<Integer>getValue(ConfigurationPropertyKey.MAX_CONNECTIONS_SIZE_PER_QUERY);
+        return SchemaMetaDataLoader.load(dataSource, maxConnectionsSizePerQuery, LogicSchemas.getInstance().getDatabaseType().getName());
     }
     
     /**
