+++ sharding-proxy/sharding-proxy-backend/src/main/java/org/apache/shardingsphere/shardingproxy/backend/schema/impl/TransparentSchema.java
@@ -21,9 +21,16 @@ import lombok.Getter;
 import org.apache.shardingsphere.api.config.sharding.ShardingRuleConfiguration;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 import org.apache.shardingsphere.shardingproxy.backend.schema.LogicSchema;
+import org.apache.shardingsphere.shardingproxy.backend.schema.LogicSchemas;
 import org.apache.shardingsphere.shardingproxy.config.yaml.YamlDataSourceParameter;
+import org.apache.shardingsphere.shardingproxy.context.ShardingProxyContext;
+import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaData;
+import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaDataLoader;
+import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.underlying.common.metadata.ShardingSphereMetaData;
+import org.apache.shardingsphere.underlying.common.metadata.datasource.DataSourceMetas;
 
+import javax.sql.DataSource;
 import java.sql.SQLException;
 import java.util.Map;
 
@@ -33,16 +40,26 @@ import java.util.Map;
 @Getter
 public final class TransparentSchema extends LogicSchema {
     
+    private final ShardingSphereMetaData metaData;
+    
     private final ShardingRule shardingRule;
     
     public TransparentSchema(final String name, final Map<String, YamlDataSourceParameter> dataSources) throws SQLException {
         super(name, dataSources);
         // TODO we should remove it after none-sharding parsingEngine completed.
         shardingRule = new ShardingRule(new ShardingRuleConfiguration(), getDataSources().keySet());
+        metaData = createMetaData();
     }
     
-    @Override
-    public ShardingSphereMetaData getMetaData() {
-        return getPhysicalMetaData();
+    private ShardingSphereMetaData createMetaData() throws SQLException {
+        DataSourceMetas dataSourceMetas = new DataSourceMetas(LogicSchemas.getInstance().getDatabaseType(), getDatabaseAccessConfigurationMap());
+        SchemaMetaData schemaMetaData = loadSchemaMetaData();
+        return new ShardingSphereMetaData(dataSourceMetas, schemaMetaData);
+    }
+    
+    private SchemaMetaData loadSchemaMetaData() throws SQLException {
+        DataSource dataSource = getBackendDataSource().getDataSources().values().iterator().next();
+        int maxConnectionsSizePerQuery = ShardingProxyContext.getInstance().getProperties().<Integer>getValue(ConfigurationPropertyKey.MAX_CONNECTIONS_SIZE_PER_QUERY);
+        return SchemaMetaDataLoader.load(dataSource, maxConnectionsSizePerQuery, LogicSchemas.getInstance().getDatabaseType().getName());
     }
 }
